Refactoring part of how argument of periapsis maneuvers are implemented, adding some tests
To account for varying orbital parameters by the time an argument of periapsis change maneuver is actually implemented (compared to initial orbital elements of the satellite), and to try to streamline the user experience more, the maneuver is simply added to the satellite, where it is stashed in a list of maneuvers that are awaiting initialization. When the satellite's internal clock reaches or first surpasses the start time, it is initialized and removed from that list of maneuvers needing initialization. Also, the calibration method previously implemented utilizing mean values has been removed, as I no longer think it's as useful as I did earlier.
Also added a couple tests relating to argument of periapsis change maneuvers.